check for log file exists

+++ migrate_helper_scripts/archive_logs.py
@@ -20,18 +20,19 @@ def archive(command="archive", volumes=False):
     logs = list_logs.get_logs(command, volumes)
     log_total = archived = removed = 0
     for log in logs:
-        log_total += 1
         year_month = get_year_month(log)
         os.makedirs(LOG_DIRECTORY + ARCHIVE_DIR + year_month, exist_ok=True)
         log_file_path = LOG_DIRECTORY + log
-        gz_file_path = LOG_DIRECTORY + ARCHIVE_DIR + year_month + '/' + log + '.gz'
-        with open(log_file_path, 'rb') as f_in:
-            with gzip.open(gz_file_path, 'wb') as f_out:
-                shutil.copyfileobj(f_in, f_out)
-                archived += 1
-        if os.path.exists(gz_file_path):
-            os.remove(log_file_path)
-            removed += 1
+        if os.path.exists(log_file_path):
+            log_total += 1
+            gz_file_path = LOG_DIRECTORY + ARCHIVE_DIR + year_month + '/' + log + '.gz'
+            with open(log_file_path, 'rb') as f_in:
+                with gzip.open(gz_file_path, 'wb') as f_out:
+                    shutil.copyfileobj(f_in, f_out)
+                    archived += 1
+            if os.path.exists(gz_file_path):
+                os.remove(log_file_path)
+                removed += 1
 
     return {'log count': log_total, 'archived': archived, 'removed': removed}
